requirements.txt
Django==5.1.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers
daphne
channels
channels-redis
redis